+++ Podfile
@@ -1,5 +1,5 @@
 # Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
+ platform :ios, '12.0'
 
 target 'WeatherForecast' do
   # Comment the next line if you don't want to use dynamic frameworks
@@ -7,11 +7,12 @@ target 'WeatherForecast' do
 
   # Pods for WeatherForecast
   pod 'SwiftLint'
-  pod 'RealmSwift'
+  pod 'RealmSwift', '~>10'
 
   target 'WeatherForecastTests' do
     inherit! :search_paths
     # Pods for testing
+    pod 'RealmSwift', '~>10'
   end
 
   target 'WeatherForecastUITests' do
