--- WeatherForecast/LocalDatabaseService/WeatherObject.swift
@@ -1,13 +1,0 @@
-//
-//  WeatherObject.swift
-//  WeatherForecast
-//
-//  Created by httruc on 28/02/2021.
-//
-
-import Foundation
-
-struct WeatherObject {
-    let city: City
-    let weatherItems: [WeatherItem]
-}
